Conflicts:
	headpin/katello-headpin/katello-headpin.spec
	rel-eng/packages/katello-headpin

+++ headpin/katello-headpin/gen_changes.sh
@@ -1,13 +1,15 @@
 #!/bin/bash
 HEADPIN=$PWD
 CHANGES_FILE="file_changes.txt"
+BRANCH=$1
+[ $# -eq 0 ] && { echo "Please specify a branch or tag; eg $0 master" ; exit 1; }
 
 # Clear out the old things
 rm $CHANGES_FILE
 rm -rf src
 
 # generate the new file
-git diff master --name-only | grep -v "katello.spec" | grep "^src/"   > $CHANGES_FILE
+git diff $BRANCH --name-only | grep -v "katello.spec" | grep "^src/"   > $CHANGES_FILE
 
 # Copy over the changes
 cd ../..
